remove package-lock.json from prettier ignore
