+++ .prettierignore
@@ -5,4 +5,3 @@ dist
 **/android/*
 **/ios/*
 .storybook.native/story-loader.js
-**/package-lock.json
