+++ packages/tracking/__tests__/tracking-child-views.test.js
@@ -1,64 +1,13 @@
-import { Text, FlatList } from "react-native";
 import React from "react";
-import PropTypes from "prop-types";
 import renderer from "react-test-renderer";
 import Enzyme, { shallow } from "enzyme";
 import React16Adapter from "enzyme-adapter-react-16";
 import { withTrackChildViews } from "../tracking";
 import withTestContext from "./test-tracking-context";
 import sharedTrackingTests from "./shared-tracking-tests";
+import ListComponent from "./ListComponent";
 
 Enzyme.configure({ adapter: new React16Adapter() });
-
-class ListComponent extends React.Component {
-  static get propTypes() {
-    return {
-      onViewed: PropTypes.func.isRequired,
-      items: PropTypes.arrayOf(
-        PropTypes.shape({
-          someKey: PropTypes.string,
-          someValue: PropTypes.string
-        })
-      ),
-      getChildList: PropTypes.func
-    };
-  }
-  static get defaultProps() {
-    return {
-      items: [{ someKey: "1", someValue: "one" }],
-      getChildList: () => {}
-    };
-  }
-  static get someStatic() {
-    return { foo: "bar" };
-  }
-  constructor(props, context) {
-    super(props, context);
-    this.onViewableItemsChanged = this.onViewableItemsChanged.bind(this);
-    props.getChildList(props.items);
-  }
-  onViewableItemsChanged({ info }) {
-    const filtered = info.changed.filter(item => item.isViewable);
-    filtered.forEach(item => this.props.onViewed(item));
-  }
-  render() {
-    return (
-      <FlatList
-        data={this.props.items}
-        renderItem={({ item }) => <Text>Item {item.someValue}</Text>}
-        onViewableItemsChanged={this.onViewableItemsChanged}
-        keyExtractor={({ someKey }) => someKey}
-        initialNumToRender={this.props.items.length}
-        viewabilityConfig={{
-          viewAreaCoveragePercentThreshold: 100,
-          waitForInteraction: false
-        }}
-      />
-    );
-  }
-}
-
-export { ListComponent as default };
 
 describe("WithTrackChildViews", () => {
   it("tracks child views", () => {
